rename EventApiClientImpl to EventApiClient
