--- shared/src/commonMain/kotlin/com/mindera/minderapeople/apiclient/EventApiClientImpl.kt
@@ -1,27 +1,0 @@
-package com.mindera.minderapeople.apiclient
-
-import com.mindera.minderapeople.apiclient.interfaces.IEventApiClient
-import com.mindera.minderapeople.dto.EventDTO
-import io.ktor.client.call.*
-import io.ktor.client.engine.*
-import io.ktor.client.request.*
-import io.ktor.http.*
-
-class EventApiClientImpl(engine: HttpClientEngine): IEventApiClient {
-
-    private val httpClient = MinderaPeopleAPIClient(engine).httpClient
-    private val baseUrl = MinderaPeopleAPIClient.BASE_URL
-
-    override suspend fun getEventById(userId: String, eventId: String): Result<EventDTO> {
-        return try{
-            val response = httpClient.get("${baseUrl}/events/${userId}/${eventId}")
-
-            if(response.status == HttpStatusCode.OK)
-                Result.success(response.body())
-            else
-                Result.failure(Exception(response.status.description))
-        }catch (e: Exception){
-            Result.failure(e)
-        }
-    }
-}
